pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whirr"
version = "0.0.1"
description = "Real-time road sounds inference app."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Alexander Yarmolich", email = "user@example.com" }
]
requires-python = ">=3.12.0"
dependencies = [
    "click>=8.2.1",
    "librosa>=0.11.0",
    "onnxruntime>=1.22.0",
    "python-dotenv>=1.1.0",
    "sounddevice>=0.5.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "snakeviz>=2.2.2",
    "tensorflow>=2.19.0",
    "tf2onnx>=1.16.1",
]

[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
Issues = "https://github.com/me/spam/issues"
Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
whirr = "control:cli"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "PL",   # pylint
    "RUF",  # ruff-specific rules
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit

    "I",    # isort
    "N",    # pep8-naming
    "ICN",  # flake8-import-conventions
    "PT",   # flake8-pytest-style
]
